+++ js/colorpicker.js
@@ -1,0 +1,11 @@
+/**
+ * Loads the WordPress wpColorPicker
+ *
+ * @package Extensions for Leaflet Map DSGVO
+ */
+
+jQuery( document ).ready(
+	function () {
+		jQuery( ".color-picker" ).wpColorPicker();
+	}
+);